pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.16.3",
    "fastapi>=0.116.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "yookassa>=3.5.0",
]
